# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: antsxdp2.0 admin  接口
  version: 2.0.0
  title: antsxdp 2.0 -admin
basePath: /antsxdp
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
  /captcha.jpg:
    get:
      tags:
        - 管理登录
      summary: 获取验证码
      produces:
        - application/octet-stream
      parameters:
        - name: uuid
          description: UUID
          in: query
          type: string
          required: true
  /sys-user/check:
    get:
      tags:
        - 管理登录检测
      summary: 管理登录
      produces:
        - application/json
      parameters:
        - name: userName
          description: userName
          in: query
          type: string
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Login'
  /sys/login:
    post:
      tags:
        - 管理登录
      summary: 管理登录
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/LoginForm'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Login'
  /sys/appkey/login:
    post:
      tags:
        - 管理登录
      summary: 管理APPKEY登录
      produces:
        - application/json
      parameters:
        - name: body
          description: map
          in: body
          type: string
          required: true
          default: "{\"username\":\"test\",\"password\":\"123456\",\"appKey\":\"AAABBBCCC\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/Login'
  /sys/google-auth/login:
    post:
      tags:
        - 管理登录
      summary: 管理google-auth登录
      produces:
        - application/json
      parameters:
        - name: body
          description: map
          in: body
          type: string
          required: true
          default: "{\"user\":\"test\",\"code\":\"123456\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/Login'
...
  /sys/dnsapi/list:
    post:
      tags:
        - DNS 接口管理
      summary: 获取分页列表
      produces:
        - application/json
      parameters:
        - name: body
          description: "获取DNS API 列表"
          in: body
          required: true
          default: "{\"page\":1,\"limit\":10,\"userType\":\"\",\"source\":\"\",\"user\":\"\"}"
          type: string
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dnsapi/save:
    post:
      tags:
        - DNS 接口管理
      summary: 保存DNS 接口
      produces:
        - application/json
      parameters:
        - name: body
          description: "保存DNS 接口"
          in: body
          required: true
          default: "{\"id\":0,\"userType\":1,\"userId\":1,\"remark\":\"aliyun\",\"source\":\"aliyun\",\"appDomain\":\"doo.com\",\"appId\":\"123\",\"appKey\":\"123\",\"appUrl\":\"kUrl\",\"status\":1}"
          type: string
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dnsapi/delete:
    post:
      tags:
        - DNS 接口管理
      summary: 删除DNS  接口
      produces:
        - application/json
      parameters:
        - name: body
          description: "删除 DNS 接口"
          in: body
          required: true
          default: "{\"ids\":\"1,2\"}"
          type: string
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/dnsapi/record/list:
    get:
      tags:
        - DNS 接口管理
      summary: API 对应域名的解析记录列表
      produces:
        - application/json
      parameters:
        - name: id
          in: query
          type: integer
          default: 3
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dnsapi/line/list:
    get:
      tags:
        - DNS 接口管理
      summary: API 对应域名的线路列表
      produces:
        - application/json
      parameters:
        - name: id
          in: query
          type: integer
          default: 3
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dnsapi/record/add:
    post:
      tags:
        - DNS 接口管理
      summary: 添加解析记录
      produces:
        - application/json
      parameters:
        - nae: body
          description: "添加解析记录"
          in: body
          required: true
          default: "{\"id\":1,\"top\":\"cdn20\",\"recordType\":\"A\",\"line\":\"default\",\"value\":\"1.1.1.1\",\"ttl\":600}"
          type: object
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dnsapi/record/remove:
    post:
      tags:
        - DNS 接口管理
      summary: 删除解析记录
      produces:
        - application/json
      parameters:
        - name: body
          description: "添加解析记录"
          in: body
          required: true
          default: "{\"id\":3,\"recordId\":\"1\"}"
          type: object
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dnsapi/record/clean:
    post:
      tags:
        - DNS 接口管理
      summary: 删除解析记录
      produces:
        - application/json
      parameters:
        - name: body
          description: "删除解析记录"
          in: body
          required: true
          default: "{\"dnsConfigId\":3,\"mainDomain\":\"1\",\"top\":\"\",\"recordType\":\"\",\"line\":\"\",\"value\":\"\",\"ttl\":\"\"}"
          type: object
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dnsapi/record/modify:
    post:
      tags:
        - DNS 接口管理
      summary: 修改解析记录
      produces:
        - application/json
      parameters:
        - name: body
          description: "添加解析记录"
          in: body
          required: true
          default: "{\"id\":1,\"recordId\":\"112233\",\"top\":\"cdn20\",\"recordType\":\"A\",\"line\":\"default\",\"value\":\"1.1.1.1\",\"ttl\":600}"
          type: object
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dnsapi/record/info:
    post:
      tags:
        - DNS 接口管理
      summary: 根据参数获取解析记录
      produces:
        - application/json
      parameters:
        - name: body
          description: "根据参数获取解析记录"
          in: body
          required: true
          default: "{\"id\":1,\"top\":\"cdn20\",\"recordType\":\"A\",\"line\":\"default\"}"
          type: string
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/cdnsys/ip/control/list:
    post:
      tags:
        - IP管理
      summary: ip列表
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description: "control:-1=所有,0=未定义,1=白名单（7层）,2=7层黑名单,3=3层黑名单"
          default: "{\"parentId\":0,\"page\":1,\"limit\":10,\"control\":-1,\"ip\":\"\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/cdnsys/ip/control/save:
    post:
      tags:
        - IP管理
      summary: ip（分组） 保存与修改
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          type: string
          description:
          default: "{\"id\":0,\"parentId\":0,\"ip\":\"1.1.1.1\",\"control\":1,\"status\":0,\"remark\":\"\",,\"ips\":\"1.1.1.1,1.2.2.2\"}"
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
  Login:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        token:
          description: token
          type: string
        expire:
          description: 过期时长
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
  LoginForm:
    type: object
    properties:
      username:
        description: 用户名
        default: admin
        type: string
      password:
        description: 密码
        default: admin
        type: string
      captcha:
        description: 验证码
        default: "0000"
        type: string
      uuid:
        description: UUID
        default: "0000"
        type: string
...